refactor: extract ffmpeg fps detector from adapter
